server/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
pydantic==2.12.4
google-cloud-aiplatform==1.71.1
python-dotenv==1.2.1
python-multipart==0.0.20
vertexai==1.71.1
pymongo==4.10.1
